refactor(profiler-hub): extract read_statements_base for schema dispatch
Introduce an abstract read_statements_base owning all shared result structs,
func typedefs, and statement sets, plus a pure-virtual shared subset and
default-empty backend-specific accessors. Reparent the v3 read_statements onto
it (same SQL, now overriding the interface) so a second schema backend can slot
in behind one shared type without touching any call site.
